port: thread→pthread, font ascent fixes, 64-bit struct fixes, debug logging
ThreadClass (wwlib/thread.cpp/.h):
- Replace Win32 _beginthread/SetThreadPriority/WaitForSingleObject with
  pthreads (pthread_create/detach/cancel, sched_yield, usleep)

wingdi_coretext.cpp:
- Track gdi_ascent (= |cHeight| for negative CreateFont height) separately
  from CoreText ascent to match Windows tmAscent convention exactly
- Add bmp_height cap (PointSize*2) to prevent heap-buffer-overflow in
  Store_GDI_Char when CoreText descent exceeds the bitmap budget
- Disable sub-pixel (ClearType) smoothing so Store_GDI_Char B-channel
  alpha reads are correct (sub-pixel R≠G≠B fringes caused blurry glyphs)
- GetTextMetricsA/GetTextExtentPoint32W: use gdi_ascent, cap tmHeight

ddsfile.h: void* Surface → uint32_t + static_assert(sizeof==124)
  (void* is 8 bytes on 64-bit, breaking the on-disk DDS format)

texturethumbnail.cpp: sizeof(unsigned long)→sizeof(uint32_t) for
  date_time write, matching the uint32_t used at read time

hlod.cpp / meshmdlio.cpp: temporary debug logging for main-menu visual
  investigation (IF_TITLETRANS sub-mesh names + prelit/lighting flags)

+++ original/Code/ww3d2/texturethumbnail.cpp
@@ -403,7 +403,7 @@ void ThumbnailManagerClass::Save(bool force)
 		WW3DFormat original_format=thumb->Get_Original_Texture_Format();
 		uint32_t date_time=thumb->Get_Date_Time();
 
-		thumb_file->Write(&date_time,sizeof(unsigned long));
+		thumb_file->Write(&date_time,sizeof(uint32_t));  // must match Read(&date_time,sizeof(uint32_t)) at load time; sizeof(unsigned long)==8 on 64-bit
 		thumb_file->Write(&offset,sizeof(int));
 		thumb_file->Write(&width,sizeof(int));
 		thumb_file->Write(&height,sizeof(int));
